==> notebooks/processing/00_Download_Data.ipynb <==
# Downloading external datasets

This notebook downloads datasets that are not committed to the repositoro due to their size.
You must run this notebook in order for all the data processing notebooks to work.

import os
import shutil

from covidcaremap.data import EXTERNAL_DATA_DIR, external_data_path

DATA_DIRECTORY = EXTERNAL_DATA_DIR
DOWNLOAD_CACHE = external_data_path('download_cache')

if not os.path.exists(DOWNLOAD_CACHE):
    os.makedirs(DOWNLOAD_CACHE)

## Download HCRIS Data


# HCRIS-HOSPITAL10_PROVIDER_ID_INFO.CSV
hcris_hosp10_zip_path = external_data_path('download_cache/hosp10-reports.zip')
!wget -O $hcris_hosp10_zip_path http://downloads.cms.gov/files/hcris/hosp10-reports.zip
hcris_hosp10_zip_path = external_data_path('download_cache/hosp10-reports.zip')      
!cd $DOWNLOAD_CACHE && unzip -o hosp10-reports.zip
!cd $DOWNLOAD_CACHE && mv HOSPITAL10_PROVIDER_ID_INFO.CSV  ../HCRIS-HOSPITAL10_PROVIDER_ID_INFO.CSV

# HCRIS-HCRIS_DataDictionary.csv 
hcris_dd_zip_path = external_data_path('download_cache/hospital2010-documentation.zip')
!wget -O $hcris_dd_zip_path https://www.cms.gov/files/zip/hospital2010-documentation.zip
!cd $DOWNLOAD_CACHE && unzip -o hospital2010-documentation.zip    
!cd $DOWNLOAD_CACHE && mv HCRIS_DataDictionary.CSV  ../HCRIS-HCRIS_DataDictionary.csv 

# HCRIS-hosp10_2018_RPT.CSV and HCRIS-hosp10_2018_NMRC.CSV
hcris_rpt_zip_path = external_data_path('download_cache/HOSP10FY2018.zip')
!wget -O $hcris_rpt_zip_path http://downloads.cms.gov/Files/hcris/HOSP10FY2018.zip
!cd $DOWNLOAD_CACHE && unzip -o HOSP10FY2018.zip
!cd $DOWNLOAD_CACHE && mv hosp10_2018_RPT.CSV  ../HCRIS-hosp10_2018_RPT.CSV 
!cd $DOWNLOAD_CACHE && mv hosp10_2018_NMRC.CSV  ../HCRIS-hosp10_2018_NMRC.CSV 


## Download US Census data by county.

# cc-est2018-alldata.csv
census_path = external_data_path('us-census-cc-est2018-alldata.csv')
!wget -O $census_path https://www2.census.gov/programs-surveys/popest/datasets/2010-2018/counties/asrh/cc-est2018-alldata.csv
