==> notebooks/processing/03_Merge_Facility_Information.ipynb <==
%load_ext autoreload
%autoreload 2

# Merge facility information

Merge facility data from HIFLD (Homeland Infrastructure Foundation-Level Data), HCRIS (Healthcare Cost Reporting Information System) and DH (Definitive Healthcare) datasets.

import os
import json
from os.path import join, isdir

import pandas as pd
import geopandas as gpd
import numpy as np

from covidcaremap.data import (processed_data_path, 
                               external_data_path,
                               local_data_path)
from covidcaremap.merge import match_facilities, FacilityColumns
from covidcaremap.mapping import map_facility_match_result

## Overview

This notebook performs matches between different hospital facility datasets.

See logic in `covidcaremap.merge.match_facilities` for information about how the matching is performed.

Generally, the algorithm is as follows:

- Compute the KNN (n=10) between all facilities.
- Create a graph for containing every facility in HIFLD and an edge for each of its neighbors if the distance between the two is less than `MAX_DISTANCE`. 
- Get the [connected components](https://en.wikipedia.org/wiki/Component_(graph_theory)) of that graph as a set of potentially matched facilities and pass it to a method that:
  - Determines the feasability of a match between each HIFLD facility and any DH and HCRIS facilities in the set, based on the numeric address number matching between the two or a name match. If it's deemed feasible (see `coidccaremap.merge.reduce_matched_facility_records` for exact logic), create a score between the facilities based on a [rapidfuzz](https://github.com/rhasspy/rapidfuzz) fuzz ratio for the name and address of the facilities.
  - Generate the final match set per HIFLD facility by ordering the potential matches between HIFLD and DH or HCRIS facilities, choosing the first of each of DH and HCRIS, and ensuring there's no duplicate matches.
- The matched sets over all components represents the matched facilities.

This approach also takes into account a manual override file which can force matches between the datasets.

## Code

These keys to refer to the facility datasets by.

HIFLD = 'hifld'
DH = 'dh'
HCRIS = 'hcris'

`MAX_DISTANCE` determines the maximum distance two facilities can be apart from each other and still considered as a potential match.

MAX_DISTANCE = 500 # meters

`AUTHORITATIVE_DATASET` is the dataset that all the other facility datasets match against. All facilities in this dataset will be included in the final output; any unmatched facilities in the other datasets will be dropped.

AUTHORITATIVE_DATASET = HIFLD

Read in the facility datasets and make any necessary data transformations.

hcris = gpd.read_file(processed_data_path('usa_hospital_beds_hcris2018_jsf.geojson'), encoding='utf-8')
hifld = gpd.read_file(processed_data_path('hifld_facility_data.geojson'), encoding='utf-8')
dh = gpd.read_file(processed_data_path('dh_geocoded_v2_04072020.geojson'), encoding='utf-8')

# Fix DH type weirdness
dh['NUM_ICU_BE'] = dh['NUM_ICU_BE'].astype(float)
dh['BED_UTILIZ'] = dh['BED_UTILIZ'].astype(float)

# Drop OBJECTID as that is the ID column name 
hifld = hifld.drop(columns=['OBJECTID'])

# Use a combined address field for DH as it's street address is split between two fields.
dh['addr2'] = dh['HQ_ADDRE_1'].fillna('')
dh['combined_address'] = dh['Street_Addr'] + ' ' + dh['addr2']

Read in the manual override file to force matches that were manually verified. Read in with dtype as object as not to read the IDs as float values.

manual_matches_df = pd.read_csv(external_data_path('hifld-dh-hcris-manual-matches.csv'),
                                dtype=object)

# Fill NaN values with an empty string
manual_matches_df = manual_matches_df.fillna('')

# Reformat the ID columns to match the datasets
manual_matches_df = manual_matches_df.rename(columns={
    'hifld_id': 'ID',
    'dh_id': 'OBJECTID',
    'hcris_id': 'Provider Number'
})

# zfill HCRIS ID to always be 6 digits with leading 0's.
def zfill_hcris(x):
    if x:
        return str(x).zfill(6)
    else:
        return x
manual_matches_df['Provider Number'] = manual_matches_df['Provider Number'].apply(zfill_hcris)


The configuration for the matching algorithm. It describes the dataframes and the column names of each facility dataset.

facility_datasets = {
    HIFLD: {
        'df': hifld,
        'columns': FacilityColumns(facility_id='ID',
                                   facility_name='NAME',
                                   street_address='ADDRESS')
    },
    DH: {
        'df': dh,
        'columns': FacilityColumns(facility_id='OBJECTID',
                                   facility_name='HOSP10_Name',
                                   street_address='combined_address')
    },
    HCRIS: {
        'df': hcris,
        'columns': FacilityColumns(facility_id='Provider Number',
                                   facility_name='HOSP10_Name',
                                   street_address='Street_Addr')
    }
}


Perform the matching. This can take a bit.

match_result = match_facilities(facility_datasets, 
                                authoritative_dataset='hifld',
                                manual_matches_df=manual_matches_df,
                                max_distance=MAX_DISTANCE)

Validate that all manual matches are made in the dataset

for _, row in match_result.matches.merge(manual_matches_df, 
                                         on='ID', 
                                         suffixes=['_match', '_manual']).iterrows():
    if row['OBJECTID_manual']:
        assert row['OBJECTID_match'] == row['OBJECTID_manual']
        
    if row['Provider Number_manual']:
        assert row['Provider Number_match'] == row['Provider Number_manual']


Save off a map of the match results for inspection.

map_dir = local_data_path('merge_facility_validation_maps')
if not os.path.isdir(map_dir):
    os.makedirs(map_dir)
all_map = map_facility_match_result(match_result, facility_datasets, 'hifld')
all_map.add_layer_selector()
all_map.save(os.path.join(map_dir, 
                          '{}.html'.format('-'.join([HIFLD, DH, HCRIS]))))

authoritative_dataset = 'hifld'

Save the merged facility information to GeoJSON.

match_result.merged_df.to_file(processed_data_path('hifld-dh-hcris-merged.geojson'), 
                               encoding='utf-8', 
                               driver='GeoJSON')

Save a JSON object describing all unmatched facilities.

with open(processed_data_path('unmatched-facilities_per_dataset.json'), 'w') as f:
    # Sort keys to ensure the same results have the same output.
    f.write(json.dumps(match_result.get_unmatched_dict(), indent=2, sort_keys=True))
